Stable new EEPROM version
First stable version of modified EEPROM library.

+++ EEPROM.h
@@ -6,8 +6,8 @@
 #include "main.h"
 
 uint8_t EEL_Init(uint32_t start_address);
-uint8_t EEL_Read_Variable(uint16_t virtual_address, uint16_t *data);
-uint8_t EEL_Write_Variable(uint16_t virtual_address, uint16_t data);
+uint8_t EEL_Read_Variable(uint8_t virtual_address, uint16_t *data);
+uint8_t EEL_Write_Variable(uint8_t virtual_address, uint16_t data);
 
 #endif /* __EEPROM_H */
